requirements.txt
supervisely==6.1.56
PyYAML==5.4.1
